+++ plone/scale/tests/test_storage.py
@@ -256,4 +256,7 @@ class AnnotationStorageTests(TestCase):
 
 def test_suite():
     from unittest import defaultTestLoader
+    from warnings import filterwarnings
+    filterwarnings("error", "the 'direction' option is deprecated")
+    filterwarnings("error", "the '.*' scaling mode is deprecated")
     return defaultTestLoader.loadTestsFromName(__name__)