+++ CHANGES.rst
@@ -13,6 +13,13 @@ New features:
 
 - *add item here*
 
+Changes:
+
+- The ``mode`` argument replaces the old, now deprecated, ``direction``
+  argument. The new names are ``contain`` or ``scale-crop-to-fit`` instead of
+  ``down``, ``cover`` or ``scale-crop-to-fill`` instead of ``up`` and
+  ``scale`` instead of ``thumbnail``.
+
 Bug fixes:
 
 - Fix cleanup of image scales in py3